requirements/common.txt
cryptography>=1.8.1
oauthlib>=2.0.2
PyJWT>=1.5.0
pyOpenSSL>=17.0.0